♻️ refactor : 어드민페이지 쿠폰목록 컴포넌트 구분
